refactor: get clients from TFprovider with xplane providerConfig (#15)
This PR introduces logic to get the API clients via the TF provider.
This reduces the need to copy all the API init logic. It does require
some type casting but other than that it looks fairly straightforward.

The `clients.go` file has to jump through two hoops. First turning the
`providerConfig` into a `map[string]any`, this logic is copy/pasta from
the Crossplane provider. Second turning the `map[string]any` into the TF
`providerConfig`. It might be useful to refactor upstream so that this
logic is available through `pkg/` instead of in `internal/`.